requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
firebase-admin==6.2.0
google-api-python-client==2.97.0
scikit-learn==1.3.0
tensorflow==2.13.0
nltk==3.8.1
textblob==0.17.1
numpy==1.24.3
pandas==2.1.0
matplotlib==3.7.2
seaborn==0.12.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
deepface==0.0.79
mediapipe==0.10.3
openai==0.28.0
transformers==4.33.1
pinecone-client==2.2.2
pymongo==4.5.0
simpy==4.0.1
requests==2.31.0 